IRGen: remove ELF autolink discard gadget
The ELF autolink section was previously relying on the SHF_EXCLUDE flag
being applied through module-level assembly.  Remove this gadget and
sink the handling down into LLVM.  The latest improvements to LLVM now
cause this workaround to be insufficient as the section is explicitly
marked as SHF_ALLOC, which will implicitly negate the SHF_EXCLUDE.  As a
result, we no longer discard the metadata.  Once this is sunk into LLVM
the attribute is unnecessary.  Furthermore, due to uniquing, will
actually cause a problem as there is no uniform handling for metadata
section prefixes.

rdar://82640394