[moveOnly] Add a Builtin.move operator that is overloaded over Builtin types.
All this does is map T -> @_moveOnly T at the SIL level by using a move_value
instruction. I am just using it so I can bring up the experimental move only
checker and write tests. Since it is only builtin types, for testing purposes
one needs to only use it with classes and convert the class to a
Builtin.NativeObject using something like Builtin.unsafeCastToNativeObject. But
tests can still be written, so we are ok.