[cxx-interop] ClangDeclFinder to handle problematic codegen cases
At the moment, header defined functions do necessarily get codegened by
C++-Interop if they are invoked purely on the call path from a
constructor. What this means is that when you have inline or template
functions that are header defined and are only invoked on a call graph
path that only originates from a C++ constructor that is invoked from
Swift, that you get LLVM IR Function declares instead of Function
defines with linkonceodr. This results in link errors unless those
functions happen to be invoked or template instantiated from another TU
that gets linked in.

This patch teaches the ClangDeclFinder to look for these cases and add
them to the list of decls that get handled by HandleTopLevelDecl.