Fix an incorrect assertion involving existential formation.
This assertion was tripping when the existential involved a marker
protocol, because it was checking for the number of protocols rather
than the number of protocols that require witness tables. Correct the
assertion and add a test.

Fixes rdar://83020734.

(cherry picked from commit 71c38934d358c01a75f8fd90bb279ab78540f556)