[SILGen] NFC: Moved EndBorrowCleanup to header.
Previously the declaration and definition of EndBorrowCleanup were both
within SILGenExpr.cpp.  That prevented the usage of cleanups which end
borrow scopes within other files.  Here, the declaration is moved to
Cleanup.h.  The necessary changes are made to SILGenExpr.cpp to keep the
definition of member functions in place.