[SILGen] Test: Moved tests for lexical lifetimes.
Previously, tests for the generation of lexical borrow scopes when
-enable-experimental-lexical-lifetimes is passed had been dumped into
the file testing the creation of all borrow scopes.  Here, the file
containing tests on general borrow scopes is reverted to its original
state and a new file dedicated to lexical lifetimes is added.
Subsequent tests for SILGen's creation of lexical lifetimes will go into
that new file.