Fix async_task_priority_current priority escalation
The priority escalation doesn't work quite as I thought it did. It isn't
whichever is greater, the priority of the task or the priority of the
place awaiting the result of the task. It seems to be less reliable than
that. I've changed the test accept either a default priority or that of
the main thread.