Initial fix for the name binding issue.
The core problem is that the name shadowing implementation does not consider whether a requirement witness has a name that conflicts with another instance member's name. That normally can't happen, and would probably be caught at some earlier phase of the frontend, but if a protocol requirement has a default witness, then we accept the program even if that requirement's name clashes with an instance member's.