[canonical-inst] Do not eliminate destructures that are results of apply.
There are a bunch of passes that assume that applies with tuple results will
always have a destructure. In a certain sense the destructure is treated as part
of the apply itself.

We really should eliminate this behavior by making apply take multiple
results. Until we begin that transition though, rather than breaking a lot of
code and the SIL I am going to respect that.