[SILGen] Only owned args get lexical lifetimes.
Previously, lexical lifetimes were added for all arguments with non-none
ownership.  For guaranteed arguments, that is redundant--they are
already guaranteed to remain alive by the caller.  And thanks to the
additional borrow scopes that the inliner adds, that guarantee remains
after inlining.  Here, the added lexical lifetimes are limited to only
those arguments whose ownership is owned.